+++ ChoGGi's Library/Code/GameFunctions.lua
@@ -2194,3 +2194,74 @@ do -- DeleteAllRocks
 		)
 	end
 end -- do
+
+-- build and show a list of attachments for changing their colours
+function ChoGGi.CodeFuncs.CreateObjectListAndAttaches(obj)
+	local ChoGGi = ChoGGi
+	obj = obj or ChoGGi.ComFuncs.SelObject()
+	if not obj or obj and not obj:IsKindOf("ColorizableObject") then
+		MsgPopup(
+			302535920001105--[[Select/mouse over an object (buildings, vehicles, signs, rocky outcrops).--]],
+			3595--[[Color--]]
+		)
+		return
+	end
+
+	local ItemList = {}
+
+	-- has no Attaches so just open as is
+	if obj:GetNumAttaches() == 0 then
+		ChoGGi.CodeFuncs.ChangeObjectColour(obj)
+		return
+	else
+		ItemList[#ItemList+1] = {
+			text = string.format(" %s",obj.class),
+			value = obj.class,
+			obj = obj,
+			hint = 302535920001106--[[Change main object colours.--]],
+		}
+		-- check and add attachments
+		if obj:IsKindOf("ComponentAttach") then
+			obj:ForEachAttach("",function(a)
+				if a:IsKindOf("ColorizableObject") then
+					ItemList[#ItemList+1] = {
+						text = a.class,
+						value = a.class,
+						parentobj = obj,
+						obj = a,
+						hint = string.format("%s\n%s: %s",S[302535920001107--[[Change colours of an attached object.--]]],S[302535920000955--[[Handle--]]],a.handle),
+					}
+				end
+			end)
+		end
+		-- any attaches not attached in the traditional sense (or that GetAttaches says fuck you to)
+		local IsValid = IsValid
+		for _,attach in pairs(obj) do
+			if IsValid(attach) and attach:IsKindOf("ColorizableObject") then
+				ItemList[#ItemList+1] = {
+					text = attach.class,
+					value = attach.class,
+					parentobj = obj,
+					obj = attach,
+					hint = 302535920001107--[[Change colours of an attached object.--]],
+				}
+			end
+		end
+
+	end
+
+	local function FiredOnMenuClick(sel,dialog)
+		ChoGGi.CodeFuncs.ChangeObjectColour(sel[1].obj,sel[1].parentobj,dialog)
+	end
+
+	ChoGGi.ComFuncs.OpenInListChoice{
+		callback = function()end,
+		items = ItemList,
+		title = string.format("%s: %s",S[174--[[Color Modifier--]]],RetName(obj)),
+		hint = 302535920001108--[[Double click to open object/attachment to edit (select to flash object).--]],
+		custom_type = 1,
+		custom_func = FiredOnMenuClick,
+		select_flash = true,
+	}
+end
+